requirements.txt
#torch>=2.0.0 - cos sie pierniczy tutaj
#torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
scikit-image>=0.21.0
Pillow>=10.0.0
matplotlib>=3.8.0
jupyter>=1.0.0
lpips>=0.1.4
requests>=2.31.0
tqdm>=4.66.0
scipy>=1.11.0

